Implement Rollback
This isn't the most performant way to manage transactional data because
it requires copying the data back into the main skiplist on commit. But
it keeps everything clean, which is good.

A future optimization would be more intellegently copy the skiplist
nodes in, knowing that it's being destroy right afterwards anyway.